requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
psycopg2-binary
Werkzeug
python-dotenv


